delete shared folder
