--- src/shared/query.ts
@@ -1,6 +1,0 @@
-export type Query = Record<string, string>;
-
-export const queryParser = (query: Query) => {
-  const joined = new URLSearchParams(query).toString();
-  return joined ? `?${joined}` : "";
-};
